Move www to backup and generate directly to www.
The jeff subdomain still lives here and will still be served until we
start 301'ing it.  The old www is here as a backup in case I screwed
up: I can always copy it over www.  The new place to generate files
(as per src/run.sh) is now www.